+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - `[jest-haste-map]` Add missing `@types/graceful-fs` dependency ([#9913](https://github.com/facebook/jest/pull/9913))
+- `[jest-runtime]` Vary ESM cache by query ([#9914](https://github.com/facebook/jest/pull/9914))
 
 ### Chore & Maintenance